+++ src/components/app/UriProvider.tsx
@@ -1,7 +1,0 @@
-import React, { FC, PropsWithChildren } from 'react'
-
-import UriContext from '../../UriContext.js'
-
-export const UriProvider: FC<PropsWithChildren<{ uri: string }>> = ({ uri, children }) => {
-  return <UriContext.Provider value={uri}>{children}</UriContext.Provider>
-}